% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-new.R
\docType{package}
\name{benford.analysis.grouped}
\alias{benford.analysis.grouped}
\alias{benford.analysis.grouped-package}
\title{Stremlining Benford Analysis for data validation and forensic analytics by group.}
\description{
Modified to enable group analysis functionalities; original functions are mostly kept untouched, with slight modifications in efficiency of the code and to reduce dependencies.
Goal: new functions that enable efficient analysis and forewarning of irregularities that are clustered in certain groups
}
\references{
Alexander, J. (2009). Remarks on the use of Benford's Law. Working Paper, Case Western Reserve University, Department of Mathematics and Cognitive Science.
\cr\cr Berger, A. and Hill, T. (2011). A basic theory of Benford's Law. Probability Surveys, 8, 1-126.
\cr\cr Hill, T. (1995). A statistical derivation of the significant-digit law. Statistical Science, 10(4), 354-363.
\cr\cr Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
\cr\cr Nigrini, M. J. (2011). Forensic Analyticis: Methods and Techniques for Forensic Accounting Investigations.Wiley and Sons: New Jersey.
\cr\cr
}
