% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-new.R
\name{Benford.list.plot}
\alias{Benford.list.plot}
\title{Plots a list of Benford objects}
\usage{
Benford.list.plot(ListOfBenfordObjects, except=c("summation", "abs diff"), multiple=TRUE ,...))
}
\arguments{
\item{except}{a character vector containing what plots not to plot. For details use ?plot.Benford}

\item{multiple}{plot multiple plots on the same page}

\item{ListOfBenfordObjects}{a list of benford objects.}
}
\description{
This takes a list of \code{benford} objects and generates plots for each of the objects, while providing additional information compared to the default plot.Benford function.
}
