% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-new.R
\name{benford.select.group}
\alias{benford.select.group}
\title{Subset Benford objects based on anomalous results}
\usage{
benford.select.group(ListOfBenfordObjects, AreaThreshold = 0.1, MantissaThreshold = 0.05, LogicalRelationship = "AND")
}
\arguments{
\item{ListOfBenfordObjects}{a list of benford objects.}

\item{AreaThreshold}{this sets the *minimum* value for an object to be flagged.
This is the summed deviation of the discrete empirical distribution curve from the discrete Benford curve; in other words, this is the total area in whcih the two curves are not in agreement.}

\item{LogicalRelationship}{this takes a string - "AND", "OR" - to specify whether an object should be flagged when it satisfies one or both criteria}

\item{This}{sets the *maximum* p-value of the Mantissa arc test for an object to be flagged.}
}
\value{
A list containing objects of class Benford that satisfy one or both of the criteria.
}
\description{
This takes a list of \code{benford} objects and returns a list of \code{benford} objects that pass the specified criteria.
}
\examples{
Sublist <- benford.select.group(ListOfBenfordObjects, AreaThreshold = 0.1, MantissaThreshold = 0.05, LogicalRelationship = "AND")
plot(ListOfBenfordObjects[[1]])

}
