% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-new.R
\name{benford.sum.table}
\alias{benford.sum.table}
\title{Summary table of a list of Benford objects}
\usage{
benford.sum.table(ListOfBenfordObjects)
}
\arguments{
\item{ListOfBenfordObjects}{a list of benford objects.}
}
\description{
This takes a list of \code{benford} objects and returns a summary table of the list, containing the following for each object: name, observations, number of digits used, Mantissa statistic, p-value of MantissaStat arc test and the total deviation of the empirical distribution from the Benford distribution.
}
